Fix lack of support to ResultSet methods from CONNX Driver 4.6
Since CONNX does not support some methods of JDBC API, we
need to wrap the object to provide some alternatives so the
query can run without problems.

The problem is on get methods that takes a Calendar and all set
methods. They all throws exceptions saying that the method is
not supported.

Note that these wrappers only solves the get methods.